Login.cs - Fixed password validation (added username to the password query), previous validation did not check for password that binds to a username. Query for validating username and password is set to only specifically select that particular column for the needed information.
Maintenance.cs - Changed the access modifier of the method fetchItemCategories to public.

UpdateCategory.cs - Added a statement to call for the fetchItemCategories method of Maintenance.cs upon clicking the update category button. Ensuring that upon updating the name of a selected category, it will also refresh the contents of the combo box in Maintenance form under the Items tab, displaying the updated category.

Set the default window state of the following Window Form to Maximized:
-Maintenance
-Delivery
-Inventory Monitoring
-Point of Sales
-Sales Monitoring